Set multitap options to "disabled" by default
Due to moving input selector to frontend instead from the core options,
ports 3-8 can be always set to standard causing multitap to be auto-activated.
With multitaps enabled, some games would misbehave or inputs not working
at all for those that are not multiplayer capable.

Setting core option multitap1 and multitap2 to disable for now.