Fix for Sim Theme Park/Alundra 2 internal crash.
Sim Theme Park requests a resolution of 320x176 for the FMV videos.
However, the limit for the height was 192.
This incorrectly resulted in the code adjusting it for 192,
eventually trigerring an internal crash.

Alundra suffered from a similar issue except it was worse :
it was setting a resolution of 128x32 temporarely in between
the loading screens.
Unlike Sim Theme Park, it's a completely black screen.

It seems that the games use DISPENV.screen struct for that purpose
and the real hardware would still display it as 320x224, just centered.