GTE stalls/timings
This is to fix an issue with Battle Arena Toshinden 1 going too fast.
Zero Divide is also said to be affected by this as well.

This finally fixes https://github.com/notaz/pcsx_rearmed/issues/133.

Note that this doesn't fix it for gte_arm.S : gtestall needs to be
set here.

Co-authored-by: firnis <firnis@ya.ru>