Remove CdlGetlocP hack.
This was added back in 2013 or so in PCSX Rearmed
and according to some tests against Tomb Raider 1 (which
is affected by the GetLocP code),
it works properly without this hack.

So let's just remove it as we are now doing it properly.
(Besides, Duckstation and mednafen don't have this hack)