lightrec: Move lightrec_code_inv() out of glue code
We're not going to add a lightrec_code_inv() in there for every port of
PCSX there is. Instead, add a on/off macro that will tell the glue code
to reference an external lightrec_code_inv(), which has to be provided
externally.

The WiiU implementation has moved to libpcsxcore/lightrec/mem_wiiu.c.
The Wii/GC implementation will move to the external WiiSX project
directly.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>