Update timing (override) hack for Power Rangers: Lightspeed Rescue (SuperrSonic)
The game's jumping command does not work if the game FPS is over 30.

At 280, on level 3 has audio stutters. Needs more testing.

SuperrSonic found 310 ideal to maintain the emulation at full speed, at the cost of the game dropping more frames, this allowed jumping to always work and keep the sound stutter free, as it was tested the entire game this way.

Credit to SuperrSonic for this fix.